@@ -15,39 +15,58 @@
         "test": "echo \"Error: no test specified\" && exit 1"
     },
     "dependencies": {
+        "@babel/core": "^7.0.0",
         "babel-plugin-styled-components": "^1.10.6",
         "baffle": "^0.3.6",
-        "eslint": "^5.16.0",
-        "eslint-config-airbnb": "^16.1.0",
-        "eslint-plugin-flowtype": "^2.50.3",
+        "eslint": "^6.3.0",
+        "eslint-config-airbnb": "^18.0.1",
+        "eslint-plugin-flowtype": "^4.3.0",
         "gatsby": "^2.15.14",
         "gatsby-plugin-google-analytics": "^2.1.16",
         "gatsby-plugin-react-helmet": "^3.1.7",
-        "gatsby-plugin-styled-components": "^1.0.5",
+        "gatsby-plugin-styled-components": "^3.1.5",
         "gatsby-plugin-typography": "^2.3.7",
         "gatsby-source-filesystem": "^2.1.22",
-        "gatsby-transformer-remark": "^1.7.44",
+        "gatsby-transformer-remark": "^2.6.22",
         "glamor": "^2.20.40",
-        "glamorous": "^4.13.1",
+        "glamorous": "^5.0.0",
         "react": "^16.9.0",
         "react-dom": "^16.9.0",
         "react-helmet": "^5.2.1",
         "react-inline-svg-icon": "^0.1.2",
         "react-typography": "^0.16.19",
         "safe-access": "^0.1.0",
-        "styled-components": "^2.4.1",
+        "styled-components": "^4.3.2",
         "typescript": "^3.6.2",
         "typography": "^0.16.19",
-        "typography-theme-funston": "^0.15.10"
+        "typography-theme-funston": "^0.16.19"
     },
     "devDependencies": {
+        "@babel/core": "^7.0.0",
+        "@babel/plugin-proposal-class-properties": "^7.5.5",
+        "@babel/plugin-proposal-decorators": "^7.0.0",
+        "@babel/plugin-proposal-do-expressions": "^7.0.0",
+        "@babel/plugin-proposal-export-default-from": "^7.5.2",
+        "@babel/plugin-proposal-export-namespace-from": "^7.0.0",
+        "@babel/plugin-proposal-function-sent": "^7.0.0",
+        "@babel/plugin-proposal-json-strings": "^7.0.0",
+        "@babel/plugin-proposal-logical-assignment-operators": "^7.0.0",
+        "@babel/plugin-proposal-nullish-coalescing-operator": "^7.0.0",
+        "@babel/plugin-proposal-numeric-separator": "^7.0.0",
+        "@babel/plugin-proposal-optional-chaining": "^7.0.0",
+        "@babel/plugin-proposal-pipeline-operator": "^7.0.0",
+        "@babel/plugin-proposal-throw-expressions": "^7.0.0",
+        "@babel/plugin-syntax-dynamic-import": "^7.0.0",
+        "@babel/plugin-syntax-import-meta": "^7.0.0",
+        "@babel/preset-env": "^7.6.0",
+        "@babel/preset-react": "^7.0.0",
         "eslint-plugin-import": "^2.18.2",
         "eslint-plugin-react": "^7.14.3",
         "prettier": "^1.18.2"
     },
     "peerDependencies": {
-        "gatsby": ">=1",
         "eslint": "^4.9.0",
+        "gatsby": ">=1",
         "react": "15.x"
     }
 }
